+++ src/plonemeeting/portal/core/browser/sync.py
@@ -3,6 +3,7 @@
 from plone import api
 from plone.app.textfield.value import RichTextValue
 from plone.autoform.form import AutoExtensibleForm
+from Products.CMFCore.Expression import Expression, getExprContext
 from z3c.form import button
 from z3c.form.form import Form
 from zope import schema
@@ -44,7 +45,19 @@ def format_attendees(meeting_data):
     return RichTextValue(formated_attendees, "text/html", "text/html")
 
 
-def sync_items(to_localized_time, meeting, items_data):
+def get_decision_from_json(deliberation_tal_format, item, item_data):
+    if not deliberation_tal_format:
+        raise AttributeError(
+            "deliberation_tal_format is invalid {}".format(deliberation_tal_format)
+        )
+    expression = Expression(deliberation_tal_format)
+    expression_context = getExprContext(item)
+    expression_context.vars["json"] = item_data
+    expression_result = expression(expression_context)
+    return expression_result
+
+
+def sync_items(to_localized_time, meeting, items_data, deliberation_tal_format):
     nb_created = nb_modified = nb_deleted = 0
     existing_items_brains = api.content.find(
         context=meeting, portal_type="Item", linkedMeetingUID=meeting.UID()
@@ -79,12 +92,18 @@ def sync_items(to_localized_time, meeting, items_data):
         # Sync item fields values
         item.plonemeeting_last_modified = modification_date
         item.title = item_title
-        # TODO use formatted intem number when available
+        # TODO use formatted item number when available
         item.number = str(item_data.get("itemNumber") / 100.0)
         item.representatives_in_charge = item_data.get("groupsInCharge")
-        # TODO item.deliberation (with tal formatting)
+
+        item.deliberation = get_decision_from_json(
+            deliberation_tal_format, item, item_data
+        )
+
         item.item_type = item_data.get("listType")
-        item.category = get_global_category(meeting.aq_parent, item_data.get("category"))
+        item.category = get_global_category(
+            meeting.aq_parent, item_data.get("category")
+        )
         item.reindexObject()
         if created:
             nb_created += 1
@@ -116,7 +135,9 @@ def sync_meeting(to_localized_time, institution, meeting_data):
         meeting.plonemeeting_uid = meeting_UID
     else:
         meeting = brains[0].getObject()
-    meeting.plonemeeting_last_modified = meeting_data.get("modification_date")
+    modification_date_str = meeting_data.get("modification_date")
+    localized_time = to_localized_time(modification_date_str, long_format=1)
+    meeting.plonemeeting_last_modified = dateutil.parser.parse(localized_time)
     meeting.attendees = format_attendees(meeting_data)
     meeting.title = meeting_title
     localized_time = to_localized_time(meeting_date, long_format=1)
@@ -184,7 +205,12 @@ class ImportMeetingForm(AutoExtensibleForm, Form):
             return
 
         json_items = json.loads(response.text)
-        results = sync_items(to_localized_time, meeting, json_items)
+        results = sync_items(
+            to_localized_time,
+            meeting,
+            json_items,
+            institution.item_decision_formatting_tal,
+        )
 
         status_msg = _(
             u"meeting_imported",