+++ src/plonemeeting/portal/core/browser/utils.py
@@ -26,19 +26,13 @@ class UtilsView(BrowserView):
 
     def get_plonemeeting_last_modified(self):
         if self.context.plonemeeting_last_modified:
-            if self.context.portal_type == "Item":
-                if not (
-                    self.context.plonemeeting_last_modified.hour == 0
-                    and self.context.plonemeeting_last_modified.minute == 0
-                ):
-                    return self.context.plonemeeting_last_modified.strftime(
-                        "%d/%m/%Y %H:%M:%S"
-                    )
-            elif self.context.portal_type == "Meeting":
-                plonemeeting_last_modified_object = datetime.datetime.strptime(
-                    self.context.plonemeeting_last_modified, "%Y-%m-%dT%H:%M:%S+00:00"
+            if not (
+                self.context.plonemeeting_last_modified.hour == 0
+                and self.context.plonemeeting_last_modified.minute == 0
+            ):
+                return self.context.plonemeeting_last_modified.strftime(
+                    "%d/%m/%Y %H:%M:%S"
                 )
-                return plonemeeting_last_modified_object.strftime("%d/%m/%Y %H:%M:%S")
 
     @mutually_exclusive_parameters("meeting", "UID")
     def get_meeting_url(self, meeting=None, UID=None):