+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 
 1.0a3 (unreleased)
 ------------------
+
+- Add tests for utils, item and institution views
+  [mpeeters]
+
+- Add `plonemeeting_last_modified` for demo data
+  [mpeeters]
+
 - Fix attendees unicode
   [boulch]